# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ReJoEs Clothing Library – Backend API
  description: |
    API for ReJoEs clothing library POS system.
    
    **Important Notes:**
    - This API is used only by Shopify POS staff and Shopify webhooks
    - Photo-based item tracking (no SKUs required)
    - One national inventory pool shared across all stores
    - Subscription tiers enforce monthly limits
    - Swap = return + checkout in one operation
    - Monthly counters reset on billing cycle
    
    **Idempotency:**
    - `x-idempotency-key` header required for checkout, return, swap operations
    - Uploads and webhooks do not require idempotency
    - Duplicate requests with same key return cached response
  version: 1.0.0
  contact:
    name: ReJoEs API Support

servers:
  - url: http://localhost:3000
    description: Local development server

tags:
  - name: Members
    description: Member lookup and validation
  - name: Loans
    description: Loan checkout, return, and swap operations
  - name: Uploads
    description: Photo upload for item identification
  - name: Webhooks
    description: Shopify webhook handlers

components:
  schemas:
    Member:
      type: object
      properties:
        id:
          type: string
          description: Unique member identifier
        shopifyCustomerId:
          type: string
          description: Shopify customer ID
        cardToken:
          type: string
          description: Physical card token for POS scanning
        tier:
          type: string
          enum: [BASIC, PLUS, PREMIUM]
          description: Subscription tier
        status:
          type: string
          enum: [ACTIVE, PAUSED, CANCELLED]
          description: Subscription status
        cycleStart:
          type: string
          format: date-time
          description: Current billing cycle start
        cycleEnd:
          type: string
          format: date-time
          description: Current billing cycle end
        itemsUsed:
          type: integer
          description: Items checked out this month
        swapsUsed:
          type: integer
          description: Swaps used this month
        itemsOut:
          type: integer
          description: Currently checked out items
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    TierConfig:
      type: object
      properties:
        itemsPerMonth:
          type: integer
          description: Maximum items allowed per month
          example: 5
        swaps:
          type: integer
          description: Maximum swaps allowed per month
          example: 2
        maxItemsOut:
          type: integer
          description: Maximum items allowed simultaneously
          example: 2

    ActiveLoanThumbnail:
      type: object
      properties:
        id:
          type: string
          description: Loan ID
        thumbnailUrl:
          type: string
          description: Thumbnail URL for POS display

    MemberResponse:
      type: object
      properties:
        member:
          $ref: '#/components/schemas/Member'
        allowances:
          $ref: '#/components/schemas/TierConfig'
        activeLoans:
          type: array
          items:
            $ref: '#/components/schemas/ActiveLoanThumbnail'

    Loan:
      type: object
      properties:
        id:
          type: string
          description: Unique loan identifier
        memberId:
          type: string
          description: Member ID
        storeLocation:
          type: string
          description: Store location identifier
        photoUrl:
          type: string
          description: Full-size photo URL
        thumbnailUrl:
          type: string
          description: Thumbnail URL for POS
        checkoutAt:
          type: string
          format: date-time
          description: When item was checked out
        dueDate:
          type: string
          format: date-time
          description: Due date (30 days from checkout)
        returnedAt:
          type: string
          format: date-time
          nullable: true
          description: When item was returned (null if active)
        createdAt:
          type: string
          format: date-time

    SwapResult:
      type: object
      properties:
        returnedLoan:
          $ref: '#/components/schemas/Loan'
        newLoan:
          $ref: '#/components/schemas/Loan'

    UploadResponse:
      type: object
      properties:
        uploadId:
          type: string
          description: Upload ID for reference in checkout/swap
        photoUrl:
          type: string
          description: Full-size photo URL
        thumbnailUrl:
          type: string
          description: Thumbnail URL

    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code for programmatic handling
            message:
              type: string
              description: Human-readable error message

  headers:
    IdempotencyKey:
      description: Unique key for idempotent requests
      required: true
      schema:
        type: string
      example: "550e8400-e29b-41d4-a716-446655440000"

paths:
  /api/members/by-card/{cardToken}:
    get:
      tags: [Members]
      summary: Get member by card token
      description: |
        Look up a member by scanning their physical card at POS.
        
        Returns member details, tier allowances, and currently active loans.
        Use this to validate membership status before checkout.
      parameters:
        - name: cardToken
          in: path
          required: true
          schema:
            type: string
          description: Card token from physical card scan
          example: "CARD123456"
      responses:
        '200':
          description: Member found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberResponse'
              examples:
                success:
                  member:
                    id: "cm123abc"
                    shopifyCustomerId: "shopify_456"
                    cardToken: "CARD123456"
                    tier: "PLUS"
                    status: "ACTIVE"
                    cycleStart: "2024-01-01T00:00:00Z"
                    cycleEnd: "2024-02-01T00:00:00Z"
                    itemsUsed: 2
                    swapsUsed: 0
                    itemsOut: 1
                    createdAt: "2024-01-01T00:00:00Z"
                    updatedAt: "2024-01-15T10:30:00Z"
                  allowances:
                    itemsPerMonth: 5
                    swaps: 2
                    maxItemsOut: 2
                  activeLoans:
                    - id: "loan789"
                      thumbnailUrl: "http://localhost:3000/uploads/thumb_abc123.jpg"
        '404':
          description: Member not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: "MEMBER_NOT_FOUND"
                  message: "Member not found"

  /api/loans/checkout:
    post:
      tags: [Loans]
      summary: Checkout an item
      description: |
        Check out a new item to a member.
        
        **Business Rules:**
        - Photo replaces SKU for item identification
        - Validates member status (ACTIVE only)
        - Enforces monthly item limit
        - Enforces max items out simultaneously
        - Due date is 30 days from checkout
        
        **Requires:** `x-idempotency-key` header
      headers:
        - $ref: '#/components/headers/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [memberId, storeLocation, uploadId]
              properties:
                memberId:
                  type: string
                  description: Member ID
                storeLocation:
                  type: string
                  description: Store location identifier
                uploadId:
                  type: string
                  description: Upload ID from /api/uploads/loan-photo
            example:
              memberId: "cm123abc"
              storeLocation: "store_ny"
              uploadId: "upload_456"
      responses:
        '201':
          description: Loan created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Loan'
              example:
                id: "loan789"
                memberId: "cm123abc"
                storeLocation: "store_ny"
                photoUrl: "http://localhost:3000/uploads/photo_abc123.jpg"
                thumbnailUrl: "http://localhost:3000/uploads/thumb_abc123.jpg"
                checkoutAt: "2024-01-15T10:30:00Z"
                dueDate: "2024-02-14T10:30:00Z"
                returnedAt: null
                createdAt: "2024-01-15T10:30:00Z"
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                subscription_inactive:
                  error:
                    code: "SUBSCRIPTION_INACTIVE"
                    message: "Subscription is paused - no new loans allowed"
                monthly_limit_exceeded:
                  error:
                    code: "MONTHLY_LIMIT_EXCEEDED"
                    message: "PLUS plan: 3 of 5 items remaining this month"
                max_items_out:
                  error:
                    code: "MAX_ITEMS_OUT_REACHED"
                    message: "PLUS plan: Maximum 2 items allowed out. Return items to continue."
                invalid_upload:
                  error:
                    code: "INVALID_UPLOAD"
                    message: "Invalid upload reference"
        '404':
          description: Member or upload not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: "MEMBER_NOT_FOUND"
                  message: "Member not found"

  /api/loans/return:
    post:
      tags: [Loans]
      summary: Return an item
      description: |
        Return a checked-out item.
        
        **Business Rules:**
        - Validates loan belongs to member
        - Cannot return already-returned loan
        - Decrements itemsOut counter
        
        **Requires:** `x-idempotency-key` header
      headers:
        - $ref: '#/components/headers/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [memberId, loanId]
              properties:
                memberId:
                  type: string
                  description: Member ID
                loanId:
                  type: string
                  description: Loan ID to return
            example:
              memberId: "cm123abc"
              loanId: "loan789"
      responses:
        '200':
          description: Loan returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Loan'
              example:
                id: "loan789"
                memberId: "cm123abc"
                storeLocation: "store_ny"
                photoUrl: "http://localhost:3000/uploads/photo_abc123.jpg"
                thumbnailUrl: "http://localhost:3000/uploads/thumb_abc123.jpg"
                checkoutAt: "2024-01-15T10:30:00Z"
                dueDate: "2024-02-14T10:30:00Z"
                returnedAt: "2024-01-20T14:00:00Z"
                createdAt: "2024-01-15T10:30:00Z"
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                already_returned:
                  error:
                    code: "LOAN_ALREADY_RETURNED"
                    message: "Loan already returned"
                wrong_member:
                  error:
                    code: "LOAN_NOT_FOUND"
                    message: "Loan does not belong to member"
        '404':
          description: Loan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: "LOAN_NOT_FOUND"
                  message: "Loan not found"

  /api/loans/swap:
    post:
      tags: [Loans]
      summary: Swap an item
      description: |
        Return one item and check out another in a single operation.
        
        **Business Rules:**
        - Swap = return + checkout in one transaction
        - Validates member has at least one item out
        - Enforces monthly swap limit
        - Enforces monthly item limit (swap counts as new checkout)
        - Due date is 30 days from swap
        
        **Requires:** `x-idempotency-key` header
      headers:
        - $ref: '#/components/headers/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [memberId, loanId, storeLocation, uploadId]
              properties:
                memberId:
                  type: string
                  description: Member ID
                loanId:
                  type: string
                  description: Loan ID to return
                storeLocation:
                  type: string
                  description: Store location identifier
                uploadId:
                  type: string
                  description: Upload ID for new item
            example:
              memberId: "cm123abc"
              loanId: "loan789"
              storeLocation: "store_ny"
              uploadId: "upload_999"
      responses:
        '200':
          description: Swap completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwapResult'
              example:
                returnedLoan:
                  id: "loan789"
                  memberId: "cm123abc"
                  storeLocation: "store_ny"
                  photoUrl: "http://localhost:3000/uploads/photo_abc123.jpg"
                  thumbnailUrl: "http://localhost:3000/uploads/thumb_abc123.jpg"
                  checkoutAt: "2024-01-15T10:30:00Z"
                  dueDate: "2024-02-14T10:30:00Z"
                  returnedAt: "2024-01-20T14:00:00Z"
                  createdAt: "2024-01-15T10:30:00Z"
                newLoan:
                  id: "loan999"
                  memberId: "cm123abc"
                  storeLocation: "store_ny"
                  photoUrl: "http://localhost:3000/uploads/photo_xyz789.jpg"
                  thumbnailUrl: "http://localhost:3000/uploads/thumb_xyz789.jpg"
                  checkoutAt: "2024-01-20T14:00:00Z"
                  dueDate: "2024-02-19T14:00:00Z"
                  returnedAt: null
                  createdAt: "2024-01-20T14:00:00Z"
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                swap_limit_exceeded:
                  error:
                    code: "SWAP_LIMIT_EXCEEDED"
                    message: "PLUS plan: 2 of 2 swaps remaining this month"
                no_items_out:
                  error:
                    code: "NO_ITEMS_TO_SWAP"
                    message: "No items out to swap"
                monthly_limit_reached:
                  error:
                    code: "MONTHLY_LIMIT_EXCEEDED"
                    message: "PLUS plan: Monthly item limit reached"
        '404':
          description: Loan or upload not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: "LOAN_NOT_FOUND"
                  message: "Loan not found"

  /api/loans/active/{memberId}:
    get:
      tags: [Loans]
      summary: Get active loans for member
      description: |
        List all currently checked-out items for a member.
        
        Returns loans sorted by checkout date (newest first).
        Use this to display member's current inventory at POS.
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
          description: Member ID
          example: "cm123abc"
      responses:
        '200':
          description: Active loans retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Loan'
              example:
                - id: "loan789"
                  memberId: "cm123abc"
                  storeLocation: "store_ny"
                  photoUrl: "http://localhost:3000/uploads/photo_abc123.jpg"
                  thumbnailUrl: "http://localhost:3000/uploads/thumb_abc123.jpg"
                  checkoutAt: "2024-01-15T10:30:00Z"
                  dueDate: "2024-02-14T10:30:00Z"
                  returnedAt: null
                  createdAt: "2024-01-15T10:30:00Z"

  /api/uploads/loan-photo:
    post:
      tags: [Uploads]
      summary: Upload item photo
      description: |
        Upload a photo of an item for checkout.
        
        **Important:**
        - Photo replaces SKU for item identification
        - Returns uploadId to reference in checkout/swap
        - Photo is automatically deleted when used in a loan
        - Max file size: 5MB
        - Supported formats: JPEG, PNG
        
        **No idempotency required** (uploads are idempotent by design)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [photo]
              properties:
                photo:
                  type: string
                  format: binary
                  description: Photo file (max 5MB)
      responses:
        '201':
          description: Photo uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
              example:
                uploadId: "upload_456"
                photoUrl: "http://localhost:3000/uploads/photo_abc123.jpg"
                thumbnailUrl: "http://localhost:3000/uploads/thumb_abc123.jpg"
        '400':
          description: Invalid upload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: "INVALID_UPLOAD"
                  message: "Photo file is required"

  /api/webhooks/subscription:
    post:
      tags: [Webhooks]
      summary: Handle Shopify subscription webhook
      description: |
        Process Shopify subscription events (created, updated, cancelled).
        
        **Security:** Requires Shopify HMAC signature verification.
        **No idempotency required** (Shopify handles deduplication).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [subscription_created, subscription_updated, subscription_cancelled]
                  description: Event type
                data:
                  type: object
                  properties:
                    shopifyCustomerId:
                      type: string
                    cardToken:
                      type: string
                    planHandle:
                      type: string
                      enum: [basic, plus, premium]
                    status:
                      type: string
                      enum: [active, paused, cancelled]
                    cycleStart:
                      type: string
                      format: date-time
                    cycleEnd:
                      type: string
                      format: date-time
            example:
              type: "subscription_created"
              data:
                shopifyCustomerId: "shopify_456"
                cardToken: "CARD123456"
                planHandle: "plus"
                status: "active"
                cycleStart: "2024-01-01T00:00:00Z"
                cycleEnd: "2024-02-01T00:00:00Z"
      responses:
        '200':
          description: Webhook processed
          content:
            application/json:
              example:
                status: "ok"
        '401':
          description: Invalid webhook signature
          content:
            application/json:
              example:
                message: "Invalid webhook signature"

  /api/webhooks/customers/delete:
    post:
      tags: [Webhooks]
      summary: Handle Shopify customer delete webhook
      description: |
        Delete member data when Shopify customer is deleted.
        
        **Security:** Requires Shopify HMAC signature verification.
        **No idempotency required** (Shopify handles deduplication).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shopifyCustomerId:
                  type: string
                  description: Shopify customer ID to delete
            example:
              shopifyCustomerId: "shopify_456"
      responses:
        '200':
          description: Customer deleted
          content:
            application/json:
              example:
                status: "ok"
        '401':
          description: Invalid webhook signature
          content:
            application/json:
              example:
                message: "Invalid webhook signature"